separate data from table
